+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: Add `Spinner` component.
+- Feature: Create `ProgressRing` component.
 
 ## [2.60.2] - 2020-03-26